+++ secure-mobile-push/smp-android-client/securemobilepush/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true">
         <receiver
-            android:name=".SmsReceiver"
+            android:name=".sms.SmsReceiver"
             android:permission="android.permission.BROADCAST_SMS">
             <intent-filter>
                 <action android:name="android.provider.Telephony.SMS_RECEIVED" />
