+++ secure-mobile-push/smp-android-client/securemobilepush/src/main/java/de/adorsys/android/securemobilepush/sms/SmsTool.java
@@ -1,4 +1,4 @@
-package de.adorsys.android.securemobilepush;
+package de.adorsys.android.securemobilepush.sms;
 
 
 import android.Manifest;
